// src/api.d.ts
export interface components {
    schemas: {
// ...
        "Model.Ref": {
            id: string;
            providerID: string;
            variant?: string;
        };
// ...
        "Agent.Info": {
            id: string;
            name: string;
            model?: components["schemas"]["Model.Ref"];
            request: components["schemas"]["Provider.Request"];
            system?: string;
            description?: string;
            /** @enum {string} */
            mode: "subagent" | "primary" | "all";
            hidden: boolean;
            color?: components["schemas"]["Agent.Color"];
            steps?: number;
            permissions: components["schemas"]["Permission.Ruleset"];
        };
// ...
        "Session.Message.Shell": {
            id: string;
            metadata?: Record<string, never>;
            time: {
                created: number;
                completed?: number;
            };
// ...
        };
// ...
        "Model.Info": {
            id: string;
            modelID: string;
            providerID: string;
            canonical?: string;
            family?: string;
            name: string;
            compatibility?: components["schemas"]["Model.Compatibility"];
            package?: string;
            settings?: components["schemas"]["Model.Settings"];
            headers?: {
                [key: string]: string;
            };
            body?: Record<string, never>;
            capabilities: components["schemas"]["Model.Capabilities"];
            variants: components["schemas"]["Model.Variant"][];
            time: {
                released: number;
            };
            cost: components["schemas"]["Model.Cost"][];
            /** @enum {string} */
            status: "alpha" | "beta" | "deprecated" | "active";
            enabled: boolean;
            limit: {
                context: number;
                input?: number;
                output: number;
            };
        };
// ...
        "Command.Info": {
            name: string;
            description?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}